rename : index.ts -> preview.ts로 파일이름변경
